fix build error, remove compilation warnings

+++ src/odbc.cpp
@@ -17,7 +17,7 @@
 
 #include <napi.h>
 #include <time.h>
-#include <stdlib.h> 
+#include <stdlib.h>
 
 #include "odbc.h"
 #include "odbc_connection.h"
@@ -35,14 +35,14 @@ Napi::Value ODBC::Init(Napi::Env env, Napi::Object exports) {
   hEnv = NULL;
   Napi::HandleScope scope(env);
 
-  // Wrap ODBC constants in an object that we can then expand 
+  // Wrap ODBC constants in an object that we can then expand
   std::vector<Napi::PropertyDescriptor> ODBC_CONSTANTS;
 
   ODBC_CONSTANTS.push_back(Napi::PropertyDescriptor::Value("ODBCVER", Napi::Number::New(env, ODBCVER), napi_enumerable));
 
   ODBC_CONSTANTS.push_back(Napi::PropertyDescriptor::Value("SQL_COMMIT", Napi::Number::New(env, SQL_COMMIT), napi_enumerable));
   ODBC_CONSTANTS.push_back(Napi::PropertyDescriptor::Value("SQL_ROLLBACK", Napi::Number::New(env, SQL_ROLLBACK), napi_enumerable));
-  
+
   ODBC_CONSTANTS.push_back(Napi::PropertyDescriptor::Value("SQL_USER_NAME", Napi::Number::New(env, SQL_USER_NAME), napi_enumerable));
   ODBC_CONSTANTS.push_back(Napi::PropertyDescriptor::Value("SQL_PARAM_INPUT", Napi::Number::New(env, SQL_PARAM_INPUT), napi_enumerable));
   ODBC_CONSTANTS.push_back(Napi::PropertyDescriptor::Value("SQL_PARAM_INPUT_OUTPUT", Napi::Number::New(env, SQL_PARAM_INPUT_OUTPUT), napi_enumerable));
@@ -98,7 +98,7 @@ Napi::Value ODBC::Init(Napi::Env env, Napi::Object exports) {
 
   exports.Set("connect", Napi::Function::New(env, ODBC::Connect));
   exports.Set("connectSync", Napi::Function::New(env, ODBC::ConnectSync));
-  
+
   // Initialize the cross platform mutex provided by libuv
   uv_mutex_init(&ODBC::g_odbcMutex);
 
@@ -106,13 +106,13 @@ Napi::Value ODBC::Init(Napi::Env env, Napi::Object exports) {
   // Initialize the Environment handle
   SQLRETURN sqlReturnCode = SQLAllocHandle(SQL_HANDLE_ENV, SQL_NULL_HANDLE, &hEnv);
   uv_mutex_unlock(&ODBC::g_odbcMutex);
-  
+
   if (!SQL_SUCCEEDED(sqlReturnCode)) {
     DEBUG_PRINTF("ODBC::New - ERROR ALLOCATING ENV HANDLE!!\n");
     Napi::Error(env, Napi::String::New(env, ODBC::GetSQLError(SQL_HANDLE_ENV, hEnv))).ThrowAsJavaScriptException();
     return env.Null();
   }
-  
+
   // Use ODBC 3.x behavior
   SQLSetEnvAttr(hEnv, SQL_ATTR_ODBC_VERSION, (SQLPOINTER) SQL_OV_ODBC3, SQL_IS_UINTEGER);
 
@@ -121,9 +121,9 @@ Napi::Value ODBC::Init(Napi::Env env, Napi::Object exports) {
 
 ODBC::~ODBC() {
   DEBUG_PRINTF("ODBC::~ODBC\n");
-  
+
   uv_mutex_lock(&ODBC::g_odbcMutex);
-  
+
   if (hEnv) {
     SQLFreeHandle(SQL_HANDLE_ENV, hEnv);
     hEnv = NULL;
@@ -146,7 +146,7 @@ class ConnectAsyncWorker : public Napi::AsyncWorker {
     ~ConnectAsyncWorker() {}
 
   private:
-  
+
     SQLTCHAR *connectionStringPtr;
     int count;
     SQLHENV hEnv;
@@ -178,7 +178,7 @@ class ConnectAsyncWorker : public Napi::AsyncWorker {
             (SQLPOINTER) size_t(connectTimeout), // ValuePtr
             SQL_IS_UINTEGER);                    // StringLength
         }
-        
+
         if (loginTimeout > 0) {
           sqlReturnCode = SQLSetConnectAttr(
             hDBC,                              // ConnectionHandle
@@ -197,28 +197,28 @@ class ConnectAsyncWorker : public Napi::AsyncWorker {
           0,                    // BufferLength - in characters
           NULL,                 // StringLength2Ptr
           SQL_DRIVER_NOPROMPT); // DriverCompletion
-        
+
         if (SQL_SUCCEEDED(sqlReturnCode)) {
 
           // odbcConnectionObject->connected = true;
 
           HSTMT hStmt;
-          
+
           //allocate a temporary statment
           sqlReturnCode = SQLAllocHandle(SQL_HANDLE_STMT, hDBC, &hStmt);
-          
+
           //try to determine if the driver can handle
           //multiple recordsets
           sqlReturnCode = SQLGetFunctions(
             hDBC,
-            SQL_API_SQLMORERESULTS, 
+            SQL_API_SQLMORERESULTS,
             &canHaveMoreResults
           );
 
           if (!SQL_SUCCEEDED(sqlReturnCode)) {
             canHaveMoreResults = 0;
           }
-          
+
           //free the handle
           sqlReturnCode = SQLFreeHandle(SQL_HANDLE_STMT, hStmt);
 
@@ -246,18 +246,18 @@ class ConnectAsyncWorker : public Napi::AsyncWorker {
         std::vector<napi_value> connectionArguments;
         connectionArguments.push_back(Napi::External<SQLHENV>::New(env, &hEnv)); // connectionArguments[0]
         connectionArguments.push_back(Napi::External<SQLHDBC>::New(env, &hDBCs[i])); // connectionArguments[1]
-        
+
         // create a new ODBCConnection object as a Napi::Value
         connections.Set(i, ODBCConnection::constructor.New(connectionArguments));
       }
 
       // pass the arguments to the callback function
       std::vector<napi_value> callbackArguments;
-      callbackArguments.push_back(env.Null());  // callbackArguments[0]  
+      callbackArguments.push_back(env.Null());  // callbackArguments[0]
       callbackArguments.push_back(connections); // callbackArguments[1]
 
       Callback().Call(callbackArguments);
-      
+
     }
 };
 
@@ -345,7 +345,7 @@ Napi::Value ODBC::ConnectSync(const Napi::CallbackInfo& info) {
 
   unsigned int connectTimeout = 5;
   unsigned int loginTimeout = 5;
-      
+
   if (connectTimeout > 0) {
     sqlReturnCode = SQLSetConnectAttr(
       hDBC,                                // ConnectionHandle
@@ -386,19 +386,19 @@ Napi::Value ODBC::ConnectSync(const Napi::CallbackInfo& info) {
   );
 
 
-  
+
   if (SQL_SUCCEEDED(sqlReturnCode)) {
 
     HSTMT hStmt;
-    
+
     //allocate a temporary statment
     sqlReturnCode = SQLAllocHandle(SQL_HANDLE_STMT, hDBC, &hStmt);
-    
+
     //try to determine if the driver can handle
     //multiple recordsets
     sqlReturnCode = SQLGetFunctions(
       hDBC,
-      SQL_API_SQLMORERESULTS, 
+      SQL_API_SQLMORERESULTS,
       &canHaveMoreResults);
 
     if (!SQL_SUCCEEDED(sqlReturnCode)) {
@@ -414,7 +414,7 @@ Napi::Value ODBC::ConnectSync(const Napi::CallbackInfo& info) {
     returnValue = env.Null();
   }
 
-  uv_mutex_unlock(&ODBC::g_odbcMutex); 
+  uv_mutex_unlock(&ODBC::g_odbcMutex);
 
   // return the SQLError
   if (!SQL_SUCCEEDED(sqlReturnCode)) {
@@ -428,7 +428,7 @@ Napi::Value ODBC::ConnectSync(const Napi::CallbackInfo& info) {
     std::vector<napi_value> connectionArguments;
     connectionArguments.push_back(Napi::External<SQLHENV>::New(env, &hEnv)); // connectionArguments[0]
     connectionArguments.push_back(Napi::External<SQLHDBC>::New(env, &hDBC)); // connectionArguments[1]
-    
+
     // create a new ODBCConnection object as a Napi::Value
     return ODBCConnection::constructor.New(connectionArguments);
   }
@@ -507,7 +507,7 @@ SQLRETURN ODBC::BindColumns(QueryData *data) {
     data->hSTMT,       // StatementHandle
     &data->columnCount // ColumnCountPtr
   );
-                        
+
   // if there was an error, set columnCount to 0 and return
   if (!SQL_SUCCEEDED(returnCode)) {
     data->columnCount = 0;
@@ -528,7 +528,7 @@ SQLRETURN ODBC::BindColumns(QueryData *data) {
       data->hSTMT,             // StatementHandle
       i + 1,                   // ColumnNumber
       column->ColumnName,      // ColumnName
-      SQL_MAX_COLUMN_NAME_LEN, // BufferLength,  
+      SQL_MAX_COLUMN_NAME_LEN, // BufferLength,
       &column->NameLength,     // NameLengthPtr,
       &column->DataType,       // DataTypePtr
       &column->ColumnSize,     // ColumnSizePtr,
@@ -596,7 +596,7 @@ SQLRETURN ODBC::BindColumns(QueryData *data) {
       case SQL_VARCHAR:
       case SQL_LONGVARCHAR:
       default:
-      
+
         maxColumnLength = (column->ColumnSize + 1) * sizeof(SQL_C_CHAR);
         targetType = SQL_C_CHAR;
         break;
@@ -647,7 +647,7 @@ SQLRETURN ODBC::FetchAll(QueryData *data) {
   }
 
   // will return either SQL_ERROR or SQL_NO_DATA
-  SQLCloseCursor(data->hSTMT); 
+  SQLCloseCursor(data->hSTMT);
   return returnCode;
 }
 
@@ -720,7 +720,7 @@ Napi::Array ODBC::ProcessDataForNapi(Napi::Env env, QueryData *data) {
   std::vector<ColumnData*> *storedRows = &data->storedRows;
   Column **columns = data->columns;
   SQLSMALLINT columnCount = data->columnCount;
-  
+
   // The rows array is the data structure that is returned from query results.
   // This array holds the records that were returned from the query as objects,
   // with the column names as the property keys on the object and the table
@@ -747,7 +747,7 @@ Napi::Array ODBC::ProcessDataForNapi(Napi::Env env, QueryData *data) {
   //                  determined from the parameters array on this object
   Napi::Array rows = Napi::Array::New(env);
 
-  // set the 'statement' property 
+  // set the 'statement' property
   if (data->sql == NULL) {
     rows.Set(Napi::String::New(env, STATEMENT), env.Null());
   } else {
@@ -762,7 +762,7 @@ Napi::Array ODBC::ProcessDataForNapi(Napi::Env env, QueryData *data) {
   rows.Set(Napi::String::New(env, RETURN), env.Undefined()); // TODO: This doesn't exist on my DBMS of choice, need to test on MSSQL Server or similar
 
   // set the 'count' property
-  rows.Set(Napi::String::New(env, COUNT), Napi::Number::New(env, data->rowCount));
+  rows.Set(Napi::String::New(env, COUNT), Napi::Number::New(env, (double)data->rowCount));
 
   // construct the array for the 'columns' property and then set
   Napi::Array napiColumns = Napi::Array::New(env);
@@ -775,7 +775,7 @@ Napi::Array ODBC::ProcessDataForNapi(Napi::Env env, QueryData *data) {
   }
   rows.Set(Napi::String::New(env, COLUMNS), napiColumns);
 
-  // iterate over all of the stored rows, 
+  // iterate over all of the stored rows,
   for (size_t i = 0; i < storedRows->size(); i++) {
 
     Napi::Object row = Napi::Object::New(env);
@@ -862,8 +862,8 @@ Napi::Array ODBC::ProcessDataForNapi(Napi::Env env, QueryData *data) {
  *        value (Requited): The value to bind
  *        inOut (Optional): the In/Out type to use, Defaults to SQL_PARAM_INPUT
  *        dataType (Optional): The data type, defaults based on the value
- *        
- *        
+ *
+ *
  */
 
 
@@ -908,7 +908,7 @@ void ODBC::StoreBindValues(Napi::Array *values, Parameter **parameters) {
     } else {
       // TODO: Throw error, don't support other types
     }
-  } 
+  }
 }
 
 SQLRETURN ODBC::DescribeParameters(SQLHSTMT hSTMT, Parameter **parameters, SQLSMALLINT parameterCount) {
@@ -922,7 +922,7 @@ SQLRETURN ODBC::DescribeParameters(SQLHSTMT hSTMT, Parameter **parameters, SQLSM
     // "Except in calls to procedures, all parameters in SQL statements are input parameters."
     parameter->InputOutputType = SQL_PARAM_INPUT;
 
-    returnCode = SQLDescribeParam(  
+    returnCode = SQLDescribeParam(
       hSTMT,                     // StatementHandle,
       i + 1,                     // ParameterNumber,
       &parameter->ParameterType, // DataTypePtr,
@@ -991,12 +991,12 @@ std::string ODBC::GetSQLError(SQLSMALLINT handleType, SQLHANDLE handle, const ch
   DEBUG_PRINTF("ODBC::GetSQLError : handleType=%i, handle=%p\n", handleType, handle);
 
   std::string errorMessageStr;
-  
+
   // Napi::Object objError = Napi::Object::New(env);
 
   int32_t i = 0;
   SQLINTEGER native;
-  
+
   SQLSMALLINT len;
   SQLINTEGER statusRecCount;
   SQLRETURN ret;
@@ -1023,27 +1023,27 @@ std::string ODBC::GetSQLError(SQLSMALLINT handleType, SQLHANDLE handle, const ch
   }
 
   errorMessageStr += "\"errors\": [";
-  
+
   for (i = 0; i < statusRecCount; i++) {
 
     DEBUG_PRINTF("ODBC::GetSQLError : calling SQLGetDiagRec; i=%i, statusRecCount=%i\n", i, statusRecCount);
-    
+
     ret = SQLGetDiagRec(
-      handleType, 
+      handleType,
       handle,
-      (SQLSMALLINT)(i + 1), 
+      (SQLSMALLINT)(i + 1),
       (SQLTCHAR *) errorSQLState,
       &native,
       (SQLTCHAR *) errorMessage,
       ERROR_MESSAGE_BUFFER_CHARS,
       &len);
-    
+
     DEBUG_PRINTF("ODBC::GetSQLError : after SQLGetDiagRec; i=%i\n", i);
 
     if (SQL_SUCCEEDED(ret)) {
       DEBUG_PRINTF("ODBC::GetSQLError : errorMessage=%s, errorSQLState=%s\n", errorMessage, errorSQLState);
 
-      
+
       if (i != 0) {
         errorMessageStr += ',';
       }
@@ -1095,11 +1095,11 @@ Napi::Object InitAll(Napi::Env env, Napi::Object exports) {
 #ifdef dynodbc
 Napi::Value ODBC::LoadODBCLibrary(const Napi::CallbackInfo& info) {
   Napi::HandleScope scope(env);
-  
+
   REQ_STR_ARG(0, js_library);
-  
+
   bool result = DynLoadODBC(*js_library);
-  
+
   return (result) ? env.True() : env.False();W
 }
 #endif
